requirements.txt
Django>=4.2,<5.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
google-generativeai>=0.8.0
requests>=2.31.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
